Added changes to ca
