+++ fele-client-service/utils/helpers.js
@@ -1,4 +1,6 @@
-const { createHash } = require('crypto')
+const { createHash, generateKeyPairSync } = require('crypto')
+const forge = require("node-forge");
+const pki = forge.pki;
 
 const sha256 = (content) => {  
     return createHash('sha256').update(content).digest('hex')
@@ -14,7 +16,60 @@ const getChannelSelector = (channelName) => {
     }
 }
 
+const getCredentialSelector = (feleUser) => {
+    return {
+        selector: {
+            feleUser: {
+                $eq: feleUser
+            }
+        }     
+    }
+}
+
+const generateCertificate = (user,attrs) => {
+    const { publicKey, privateKey } = generateKeyPairSync('rsa', {
+      modulusLength: 2048,
+      namedCurve: 'secp256k1', 
+      publicKeyEncoding: {
+        type: 'spki',
+        format: 'pem'
+      },
+      privateKeyEncoding: {
+        type: 'pkcs8',
+        format: 'pem'
+      }
+    });
+    const prKey = pki.privateKeyFromPem(privateKey);
+    const pubKey = pki.publicKeyFromPem(publicKey);
+  
+    const cert = pki.createCertificate();
+  
+    // fill the required fields
+    cert.CN = user;
+    cert.organization = attrs.orgName;
+    cert.publicKey = pubKey;
+    cert.serialNumber = "01";
+    cert.validity.notBefore = new Date();
+    cert.validity.notAfter = new Date();
+    cert.validity.notAfter.setFullYear(
+      cert.validity.notBefore.getFullYear() + 1
+    );
+    // here we set subject and issuer as the same one
+    cert.setSubject(attrs);
+    cert.setIssuer(attrs);
+  
+    cert.sign(prKey);
+    return {
+      feleUser: user,
+      certificate: pki.certificateToPem(cert),
+      publicKey: publicKey,
+      privateKey: privateKey
+    }
+  };
+
 module.exports = {
     sha256,
-    getChannelSelector
+    getChannelSelector,
+    getCredentialSelector,
+    generateCertificate
 }
